[refactoring] NavigationItemコンポーネントの責務を分離
